New release

+++ autopilot-panel/globals.cpp
@@ -66,11 +66,11 @@ void identifyAircraft(char* aircraft)
 /// <summary>
 /// Server can send us a delta rather than full data so we need to unpack it.
 /// </summary>
-void receiveDelta(char *deltaData, long deltaSize, char* simVarsPtr)
+void receiveDelta(char *deltaData, int deltaSize, char* simVarsPtr)
 {
     char* dataPtr = deltaData;
 
-    long tempDeltaSize = deltaSize;
+    int tempDeltaSize = deltaSize;
     while (deltaSize > 0) {
         DeltaDouble* deltaDouble = (DeltaDouble*)dataPtr;
         if (deltaDouble->offset & 0x10000) {
